Skip signal errors for unix channels
I'm not sure if I'm doing anything near
the right thing here.
My problem was that under Linux,
I was not receiving a signal,
but was instead having it panic
while waiting on the select.

The reason it panic'd was the select got EINTR.
Looking at the code, it appears that signal
should be handled
properly later.
So I changed it to ignore that error
and let it get handled later.